Fix #320 and #256 - Drop Froyo support, bump to latest plugins/libraries
In the process of looking to fix #477, we need to change minSdkVersion to 9 to use the android-maps-util library.  This also means increasing the Google Play Services version to the latest, which means we have to fix #320 to switch from the LocationClient to FusedLocationApi.  Also bumps the Android Gradle Plugin version and support library vresion to match our compileSdkVersion.  Also removes an unused PNG file that was flagged by the Gradle plugin as an invalid name (it includes a ".").